requirements.txt
flask
mysql-connector-python
fpdf
pandas
